rpc/jsonrpc: trim the over-engineering left by #23690
Six sites from an over-engineering review pass on #23690.

- receiptsRetiredWithHistory is inlined as `!amount.Enabled()`: both call
  sites are the second arm of a switch whose first arm already handles
  KeepAllReceiptsPruneMode, and both hold `amount` in scope.
- TestCapabilitiesFollowHistoryForASentinelRetention becomes a
  pruneGatingConfigs row, which TestCapabilitiesAgreeWithGates pins
  against the gate more tightly than the deleted test did.
- The two new receipts test files become one test next to the helpers
  they use, pinning the pre-semaphore answer and the cache together.
- The resolveLogsRange leg, and three docstrings that restate their name
  or belong in a commit message, are dropped.

No behavior change, so no new test: the existing suite covers it.
