+++ rpc/jsonrpc/erigon_receipts_test.go
@@ -501,10 +501,8 @@ func TestGetLogsByHashZeroLogReceipt(t *testing.T) {
 	assert.JSONEq(t, `[[]]`, string(encoded))
 }
 
-// TestErigonGetLogsByBlockHashRequiresACanonicalBlock pins erigon_getLogs on the same
-// by-hash resolution as eth_getLogs: a hash no header answers for is an error rather
-// than an empty log array, and a reorged-out sibling is not served the logs of the
-// canonical block at its height.
+// TestErigonGetLogsByBlockHashRequiresACanonicalBlock pins that a reorged-out sibling
+// is not served the logs of the canonical block at its height.
 func TestErigonGetLogsByBlockHashRequiresACanonicalBlock(t *testing.T) {
 	m, _, orphanedChain := rpcdaemontest.CreateTestExecModule(t)
 	api := NewErigonAPI(newBaseApiForTest(m), m.DB, nil)
