+++ rpc/jsonrpc/eth_api.go
@@ -532,10 +532,7 @@ func (api *BaseAPI) blocksFollowChainHistoryExpiry(ctx context.Context, tx kv.Tx
 // point, which tells a legacy archive from chain-history expiry where the stored prune
 // mode carries the same sentinel for both. Only a readable transaction of an early block
 // settles it: expiry keeps pre-merge headers and bodies, and the transaction segment
-// spanning the merge point reaches below it. Block data the search cannot read leaves the
-// question open and is not remembered; a settled answer is availability rather than
-// policy, so it is kept for a short TTL in both directions. One probe answers every
-// caller waiting on it, since each costs several backend reads under an open read tx.
+// spanning the merge point reaches below it.
 func (api *BaseAPI) holdsPreMergeBlockData(ctx context.Context, tx kv.Tx, mergeHeight uint64) (bool, error) {
 	for {
 		if v := api._preMergeData.Load(); v != nil && time.Since(v.at) < api._preMergeDataTTL {
@@ -778,7 +775,7 @@ func (api *BaseAPI) checkReceiptsAvailable(ctx context.Context, tx kv.Tx, block 
 	switch amount := p.ReceiptsAmount(); {
 	case amount == prune.KeepAllReceiptsPruneMode:
 		return nil
-	case receiptsRetiredWithHistory(p):
+	case !amount.Enabled():
 		return api.checkPruneHistory(ctx, tx, block)
 	default:
 		err := api.checkPruneField(tx, block, func(*prune.Mode) prune.BlockAmount { return amount }, "receipts are available")
@@ -787,14 +784,6 @@ func (api *BaseAPI) checkReceiptsAvailable(ctx context.Context, tx kv.Tx, block 
 		}
 		return api.checkPruneHistory(ctx, tx, block)
 	}
-}
-
-// receiptsRetiredWithHistory reports whether the receipt cache is retired at the history
-// cutoff: every retention that is not a window of its own, save an explicit keep-all.
-// This is the rule historyRetireCutoffs applies.
-func receiptsRetiredWithHistory(p *prune.Mode) bool {
-	amount := p.ReceiptsAmount()
-	return amount != prune.KeepAllReceiptsPruneMode && !amount.Enabled()
 }
 
 // postStateCalculated reports whether the receipts of this block carry a post state
