+++ rpc/jsonrpc/receipts/post_state_calculated_test.go
@@ -17,12 +17,20 @@
 package receipts
 
 import (
+	"context"
 	"testing"
 
+	"github.com/holiman/uint256"
 	"github.com/stretchr/testify/require"
 
+	"github.com/erigontech/erigon/common"
+	"github.com/erigontech/erigon/db/datadir"
 	"github.com/erigontech/erigon/db/dbservices"
+	"github.com/erigontech/erigon/db/kv/temporal/temporaltest"
 	"github.com/erigontech/erigon/execution/chain"
+	"github.com/erigontech/erigon/execution/stagedsync/stages"
+	"github.com/erigontech/erigon/execution/types"
+	"github.com/erigontech/erigon/p2p/protocols/eth"
 )
 
 // unreachableFrozenBlocks fails the test if the frozen block count is read. On a remote
@@ -52,4 +60,39 @@ func TestPostStateCalculated(t *testing.T) {
 		t.Parallel()
 		require.True(t, PostStateCalculated(cfg, byzantium-1, true, unreachableFrozenBlocks{t: t}))
 	})
+}
+
+// TestGetReceiptsAnswersAnEmptyBlockWithoutAnExecSlot pins that a block with no
+// transactions is answered before the execution semaphore and cached like any other:
+// it derives nothing, so a slot spent on it is a slot kept from a block that has work
+// to do, and repeating the request must not pay for that again. The semaphore is full
+// and the context already cancelled, so any path reaching the semaphore reports the
+// cancellation instead of the empty result.
+func TestGetReceiptsAnswersAnEmptyBlockWithoutAnExecSlot(t *testing.T) {
+	t.Parallel()
+
+	db := temporaltest.NewTestDB(t, datadir.New(t.TempDir()))
+	tx, err := db.BeginTemporalRw(t.Context())
+	require.NoError(t, err)
+	defer tx.Rollback()
+
+	const blockNum = uint64(1)
+	require.NoError(t, stages.SaveStageProgress(tx, stages.Execution, blockNum))
+	block := types.NewBlockWithHeader(&types.Header{Number: *uint256.NewInt(blockNum)}, nil)
+
+	g := newTestGenerator(t)
+	g.blockExecMutex = &loaderMutex[common.Hash]{}
+	g.execSem = make(chan struct{}, 1)
+	g.execSem <- struct{}{}
+
+	ctx, cancel := context.WithCancel(t.Context())
+	cancel()
+
+	got, err := g.GetReceipts(ctx, chain.TestChainBerlinConfig, tx, block, eth.ReceiptsOpts{})
+	require.NoError(t, err)
+	require.Empty(t, got)
+
+	cached, ok := g.GetCachedReceipts(t.Context(), block.Hash())
+	require.True(t, ok, "the next request must be answered by the cache")
+	require.Empty(t, cached)
 }
